feat: ajout du code python

+++ drivers/rotary_angle/rotary_angle.cpp
@@ -43,6 +43,6 @@ uint16_t RotaryAngle::readRaw() {
 long RotaryAngle::readDegrees() {
     uint16_t raw = readRaw();
     // La plage ADC est 0 - 1023, la plage angle 0 - 300 degrés
-    long degrees = ((float)raw) * 300.0f / 1023.0f;
+    long degrees = ((long)raw) * 300L / 1023L;
     return degrees;
 }
